+++ mono/svelte.config.js
@@ -1,5 +1,5 @@
 import { mdsvex } from 'mdsvex';
-import adapter from '@sveltejs/adapter-auto';
+import adapter from '@sveltejs/adapter-cloudflare';
 import { existsSync, readdirSync } from 'node:fs';
 import { fileURLToPath } from 'node:url';